requirements.txt
rdkit
deepchem
deepxde
phiflow
aizynthfinder
torch
llvmlite
numba
